API/requirements.txt
fastapi
psycopg2
uvicorn
python-dotenv